osd: dissect ECCommon and ECListener into separated header
